Manage PDFs with reverted coordinates
On almost all PDFs, the (0,0) coordinate is at the lower left.
But in some, it's at the top left, so the tuple got when calling
cropBox does not have informations in the good order.
By computing absolute distance between the top left corner and the
bottom right one, we can have the real height and width in both cases.