Make moveIntoView() work with the special page "0".
Page 0 doesn't exist, but since we show the first page on the right side, and we calculate page
indices from the left one, 0 is passed around as the current page number in a couple places.
The slider need to take this into account to correctly move back when we go to the first page.